+++ R/mutual.R
@@ -53,19 +53,19 @@ mutual_total_within_compute <- function(data, group, unit, within, base,
         M = sum(p_unit * (entropyw - entropy_cond)),
         p = first(p_within),
         H = sum(p_unit * (entropyw - entropy_cond)) / first(entropyw),
-        h_weight = first(p_within) * first(entropyw) / entropy_overall
+        ent_ratio = first(entropyw) / entropy_overall
     ), by = within]
     by_within$H <- ifelse(is.finite(by_within$H), by_within$H, 0)
 
     if (components == TRUE) {
         melt(by_within,
-             id.vars = within, measure.vars = c("M", "p", "H", "h_weight"),
+             id.vars = within, measure.vars = c("M", "p", "H", "ent_ratio"),
              variable.name = "stat", value.name = "est",
              variable.factor = FALSE)
     } else {
         # total M is the sum of weighted within-group partial M
-        M <- sum(by_within$M %*% by_within$p)
-        H <- sum(by_within$H %*% by_within$h_weight)
+        M <- sum(by_within$M * by_within$p)
+        H <- sum(by_within$H * by_within$p * by_within$ent_ratio)
         data.table(stat = c("M", "H"), est = c(M, H), stringsAsFactors = FALSE)
     }
 }
@@ -223,10 +223,11 @@ mutual_total <- function(data, group, unit, within = NULL, weight = NULL,
 #'   \code{M} is the within-category M, and \code{p} is the proportion of the category.
 #'   Multiplying \code{M} and \code{p} gives the contribution of each within-category
 #'   towards the total M.
-#'   \code{H} is the within-category H, and \code{h_weight} provides the weight.
-#'   Multiplying \code{H} and \code{h_weight} gives the contribution of each within-category
-#'   towards the total H. \code{h_weight} is defined as \code{p * EW/E}, where \code{EW} is the
-#'   within-category entropy, and \code{E} is the overall entropy.
+#'   \code{H} is the within-category H, and \code{ent_ratio} provides the entropy ratio,
+#'   defined as \code{EW/E}, where \code{EW} is the within-category entropy,
+#'   and \code{E} is the overall entropy.
+#'   Multiplying \code{H}, \code{p}, and \code{ent_ratio} gives the contribution of each within-category
+#'   towards the total H.
 #'   If \code{se} is set to \code{TRUE}, an additional column \code{se} contains
 #'   the associated bootstrapped standard errors, an additional column \code{CI} contains
 #'   the estimate confidence interval as a list column, an additional column \code{bias} contains
@@ -247,9 +248,9 @@ mutual_total <- function(data, group, unit, within = NULL, weight = NULL,
 #' mutual_total(schools_A, "race", "school", weight = "n") # M => .409
 #'
 #' # to recover the within M and H from the output, multiply
-#' # p * M and h_weight * H, respectively
+#' # p * M and p * ent_ratio * H, respectively
 #' sum(within$p * within$M) # => .326
-#' sum(within$H * within$h_weight) # => .321
+#' sum(within$p * within$ent_ratio * within$H) # => .321
 #' # compare with:
 #' mutual_total(schools00, "race", "school", within = "state", weight = "n")
 #' @import data.table
@@ -290,19 +291,19 @@ mutual_within <- function(data, group, unit, within,
 
     if (wide == TRUE) {
         f <- stats::as.formula(paste(paste(within, collapse = "+"),
-                                     "~ factor(stat, levels=c('M', 'p', 'H', 'h_weight'))"))
+                                     "~ factor(stat, levels=c('M', 'p', 'H', 'ent_ratio'))"))
         if (se == TRUE) {
             ret <- dcast(ret, f, value.var = c("est", "se", "CI", "bias"))
             names(ret) <- c(within,
-                            "M", "p", "H", "h_weight",
-                            "M_se", "p_se", "H_se", "h_weight_se",
-                            "M_CI", "p_CI", "H_CI", "h_weight_CI",
-                            "M_bias", "p_bias", "H_bias", "h_weight_bias")
+                            "M", "p", "H", "ent_ratio",
+                            "M_se", "p_se", "H_se", "ent_ratio_se",
+                            "M_CI", "p_CI", "H_CI", "ent_ratio_CI",
+                            "M_bias", "p_bias", "H_bias", "ent_ratio_bias")
             setcolorder(ret, c(within,
                                "M", "M_se", "M_CI", "M_bias",
                                "p", "p_se", "p_CI", "p_bias",
                                "H", "H_se", "H_CI", "H_bias",
-                               "h_weight", "h_weight_se", "h_weight_CI", "h_weight_bias"))
+                               "ent_ratio", "ent_ratio_se", "ent_ratio_CI", "ent_ratio_bias"))
             setattr(ret, "bootstrap", boot_ret)
         } else {
             ret <- dcast(ret, f, value.var = c("est"))